refactor: Add IBinaryLoaderArguments for use in the BinaryLoaderSerializer class instead of a concrete class
