+++ src/CoreHook.BinaryInjection/BinaryLoader/BinaryLoaderArguments.cs
@@ -1,22 +1,16 @@
-﻿using System.Text;
-
-namespace CoreHook.BinaryInjection
+﻿
+namespace CoreHook.BinaryInjection.BinaryLoader
 {
-    public class BinaryLoaderArguments
+    public class BinaryLoaderArguments : IBinaryLoaderArguments
     {
-        public bool Verbose;
+        public bool Verbose { get; set; }
 
-        public bool WaitForDebugger;
+        public bool WaitForDebugger { get; set; }
 
-        public string PayloadFileName;
+        public string PayloadFileName { get; set; }
 
-        public string CoreRootPath;
+        public string CoreRootPath { get; set; }
 
-        public string CoreLibrariesPath;
-
-        public static byte[] GetPathArray(string path, int pathLength, Encoding encoding)
-        {
-            return encoding.GetBytes(path.PadRight(pathLength, '\0'));
-        }
+        public string CoreLibrariesPath { get; set; }
     }
 }
