+++ src/CoreHook.BinaryInjection/BinaryLoader/BinaryLoaderArgumentsHelper.cs
@@ -1,0 +1,12 @@
+﻿using System.Text;
+
+namespace CoreHook.BinaryInjection.BinaryLoader
+{
+    internal static class BinaryLoaderArgumentsHelper
+    {
+        public static byte[] GetPathArray(string path, int pathLength, Encoding encoding)
+        {
+            return encoding.GetBytes(path.PadRight(pathLength, paddingChar: '\0'));
+        }
+    }
+}
