+++ src/CoreHook.BinaryInjection/BinaryLoader/IBinaryLoaderArguments.cs
@@ -1,0 +1,16 @@
+﻿
+namespace CoreHook.BinaryInjection.BinaryLoader
+{
+    public interface IBinaryLoaderArguments
+    {
+        bool Verbose { get; }
+
+        bool WaitForDebugger { get; }
+
+        string PayloadFileName { get; }
+
+        string CoreRootPath { get; }
+
+        string CoreLibrariesPath { get; }
+    }
+}
